ritza-real-world-tests/anchor/requirements.txt
anchorbrowser
playwright
python-dotenv
